Spr6, Try2

+++ yatube/posts/views.py
@@ -3,7 +3,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.views.decorators.cache import cache_page
 
 from .forms import CommentForm, PostForm
-from .models import Comment, Group, Post, User, Follow
+from .models import Group, Post, User, Follow
 from .utils import paginator
 
 
@@ -12,7 +12,6 @@ def index(request):
     posts = Post.objects.select_related('group', 'author')
     page_obj = paginator(request, posts)
     context = {
-        'posts': posts,
         'page_obj': page_obj
     }
     return render(request, 'posts/index.html', context)
@@ -34,13 +33,11 @@ def group_posts(request, slug):
 def profile(request, username):
     user = get_object_or_404(User.objects.filter(username=username))
     page_obj = paginator(request, user.posts.select_related('group', 'author'))
-    if request.user.is_authenticated and Follow.objects.filter(
-        user=request.user,
-        author=user
-    ):
-        following = True
-    else:
-        following = False
+    following = (
+        request.user.is_authenticated
+        and request.user != user
+        and Follow.objects.filter(user=request.user, author=user)
+    )
     context = {
         'author': user,
         'page_obj': page_obj,
@@ -54,8 +51,8 @@ def post_detail(request, post_id):
         Post.objects.select_related('group', 'author'),
         id=post_id
     )
-    form = CommentForm(request.POST or None)
-    comments = Comment.objects.filter(post=post)
+    form = CommentForm()
+    comments = post.comments.select_related('author')
     context = {
         'post': post,
         'form': form,
@@ -110,37 +107,19 @@ def add_comment(request, post_id):
 def follow_index(request):
     posts = Post.objects.filter(author__following__user=request.user)
     page_obj = paginator(request, posts)
-    context = {
-        'posts': posts,
-        'page_obj': page_obj
-    }
+    context = {'page_obj': page_obj}
     return render(request, 'posts/follow.html', context)
 
 
 @login_required
 def profile_follow(request, username):
-    # Подписаться на автора
     author = get_object_or_404(User, username=username)
-    if author != request.user and not Follow.objects.filter(
-        user=request.user,
-        author=author
-    ).exists():
-        Follow.objects.create(
-            user=request.user,
-            author=author
-        )
+    Follow.objects.get_or_create(user=request.user, author=author)
     return redirect('posts:profile', author)
 
 
 @login_required
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
-    if author != request.user and Follow.objects.filter(
-        user=request.user,
-        author=author
-    ).exists():
-        Follow.objects.filter(
-            user=request.user,
-            author=author
-        ).delete()
+    Follow.objects.filter(user=request.user, author=author).delete()
     return redirect('posts:profile', author)
